[REFAC] #9 로그인/로그아웃 마이페이지로 분리, 로그아웃 기능 구현
