+++ save_u/lib/pages/login_page.dart
@@ -1,0 +1,118 @@
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+
+import '../main.dart';
+import '../sevices/auth_service.dart';
+import 'my_page.dart';
+
+/// 로그인 페이지
+class LoginPage extends StatefulWidget {
+  const LoginPage({Key? key}) : super(key: key);
+
+  @override
+  State<LoginPage> createState() => _LoginPageState();
+}
+
+class _LoginPageState extends State<LoginPage> {
+  TextEditingController emailController = TextEditingController();
+  TextEditingController passwordController = TextEditingController();
+
+  @override
+  Widget build(BuildContext context) {
+    return Consumer<AuthService>(
+      builder: (context, authService, child) {
+        final user = authService.currentUser();
+        return Scaffold(
+          appBar: AppBar(title: Text("로그인")),
+          body: SingleChildScrollView(
+            padding: const EdgeInsets.all(16),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.stretch,
+              children: [
+                /// 현재 유저 로그인 상태
+                Center(
+                  child: Text(
+                    user == null
+                        ? "로그인해 주세요 🙂"
+                        : "${user.email}님이 로그인 되어 있습니다.",
+                    style: TextStyle(
+                      fontSize: 24,
+                    ),
+                  ),
+                ),
+                SizedBox(height: 32),
+
+                /// 이메일
+                TextField(
+                  controller: emailController,
+                  decoration: InputDecoration(hintText: "이메일"),
+                ),
+
+                /// 비밀번호
+                TextField(
+                  controller: passwordController,
+                  obscureText: false, // 비밀번호 안보이게
+                  decoration: InputDecoration(hintText: "비밀번호"),
+                ),
+                SizedBox(height: 32),
+
+                /// 로그인 버튼
+                ElevatedButton(
+                  child: Text("로그인", style: TextStyle(fontSize: 21)),
+                  onPressed: () {
+                    // 로그인 성공시 HomePage로 이동
+                    authService.signIn(
+                      email: emailController.text,
+                      password: passwordController.text,
+                      onSuccess: () {
+                        // 로그인 성공
+                        ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+                          content: Text("로그인 성공"),
+                        ));
+                        String userEmailInput =
+                            user == null ? "" : "${user.email}";
+                        // HomePage로 이동
+                        Navigator.pushReplacement(
+                          context,
+                          MaterialPageRoute(builder: (context) => HomePage()),
+                        );
+                      },
+                      onError: (err) {
+                        // 에러 발생
+                        ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+                          content: Text(err),
+                        ));
+                      },
+                    );
+                  },
+                ),
+
+                /// 회원가입 버튼
+                ElevatedButton(
+                  child: Text("회원가입", style: TextStyle(fontSize: 21)),
+                  onPressed: () {
+                    //회원가입
+                    authService.signUp(
+                      email: emailController.text,
+                      password: passwordController.text,
+                      onSuccess: () {
+                        ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+                          content: Text("회원가입 성공"),
+                        ));
+                      },
+                      onError: (err) {
+                        ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+                          content: Text(err),
+                        ));
+                      },
+                    );
+                  },
+                ),
+              ],
+            ),
+          ),
+        );
+      },
+    );
+  }
+}
