+++ save_u/lib/sevices/auth_service.dart
@@ -86,6 +86,7 @@ class AuthService extends ChangeNotifier {
   }
 
   void signOut() async {
-    // 로그아웃
+    await FirebaseAuth.instance.signOut();
+    notifyListeners(); // 로그인 상태 변경 알림
   }
 }
